HACAPI-HallHop/requirements.txt
# Core Flask app
Flask>=3.1.1
flask-cors>=6.0.0
flask-jwt-extended>=4.6.0
flask-limiter>=3.5.0

# HTTP requests and parsing
requests>=2.31.0
beautifulsoup4>=4.13.0
lxml>=5.0.0
urllib3>=2.0.0

# Environment config
python-dotenv>=1.1.0

# Secure deployment
gunicorn>=23.0.0

# Supabase
supabase>=2.15.1

# Redis for secure session caching
redis>=5.0.0
certifi>=2025.4.26  # For Upstash Redis SSL support

# Data encryption
cryptography>=42.0.0